+++ loader/format_report.lua
@@ -5,7 +5,7 @@ local NAME = "format_report"
 local semver = require("semver")
 
 -- Module version
-local VERSION = semver(0, 32, 0)
+local VERSION = semver(0, 33, 0)
 
 -- Returns the module version
 local function getVersion()