+++ serde/round_trip.lua
@@ -5,7 +5,7 @@ local NAME = "round_trip"
 local semver = require("semver")
 
 -- Module version
-local VERSION = semver(0, 32, 0)
+local VERSION = semver(0, 33, 0)
 
 -- Dependencies
 local read_only = require("util.read_only")