+++ serde/serialization.lua
@@ -5,7 +5,7 @@ local NAME = "serialization"
 local semver = require("semver")
 
 -- Module version
-local VERSION = semver(0, 32, 0)
+local VERSION = semver(0, 33, 0)
 
 -- Dependencies
 local sparse_seq = require("util.sparse_sequence")